chore: Move mappers to a different file
